Add Error handling middleware
